backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
python-multipart==0.0.6
openai==1.58.1
anthropic==0.42.0
markdown==3.5.2
pydantic==2.5.3
pydantic-settings==2.1.0
